split out task due date display
